# api/openapi.yaml
openapi: 3.0.1
info:
  title: FlopBox API
  description: API REST pour la gestion centralisée des fichiers sur des serveurs
    FTP tiers.
  version: 1.0.0
servers:
  - url: https://virtserver.swaggerhub.com/LucasDeloison/FlopBox/1.0.0
    description: SwaggerHub API Auto Mocking
  - url: https://api.flopbox.com/v2
    description: Serveur de production
  - url: http://localhost:8080
    description: Serveur de développement
...
paths:
...
  /v2/ftp-servers:
    get:
      tags:
        - Serveurs FTP
      summary: Lister les serveurs FTP
      operationId: listFTPServers
      responses:
        "200":
          description: Liste des serveurs FTP
      security:
        - BearerAuth: []
      x-swagger-router-controller: ServeursFTP
    post:
      tags:
        - Serveurs FTP
      summary: Ajouter un serveur FTP
      operationId: addFTPServer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FTPServer'
      responses:
        "201":
          description: Serveur FTP ajouté
      security:
        - BearerAuth: []
      x-swagger-router-controller: ServeursFTP
...
  /v2/ftp-servers/{alias}:
    get:
      tags:
        - Serveurs FTP
      summary: Récupérer un serveur FTP par alias
      operationId: getFTPServer
      parameters:
        - name: alias
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: Informations du serveur FTP
      security:
        - BearerAuth: []
      x-swagger-router-controller: ServeursFTP
    put:
      tags:
        - Serveurs FTP
      summary: Mettre à jour un serveur FTP
      operationId: updateFTPServer
      parameters:
        - name: alias
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FTPServer'
      responses:
        "200":
          description: Serveur FTP mis à jour
      security:
        - BearerAuth: []
      x-swagger-router-controller: ServeursFTP
    delete:
      tags:
        - Serveurs FTP
      summary: Supprimer un serveur FTP
      operationId: deleteFTPServer
      parameters:
        - name: alias
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "204":
          description: Serveur supprimé
      security:
        - BearerAuth: []
      x-swagger-router-controller: ServeursFTP
...
components:
  schemas:
...
    FTPServer:
      type: object
      properties:
        alias:
          type: string
        hostname:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT